stream_data/requirements.txt
pandas
redis
python-dotenv
requests
websocket-client
